Replace MutuallyExclusive component with MultipleChoiceAnswer
